Move dockerimages -> image_remote_lookup
better explains what it does and less likey to be confused with the
images.py module I'm working on

Also move the job submit code to be in the same order as the routers. No
changes otherwise
